Change types to reduce bytecount

+++ game.c
@@ -12,7 +12,7 @@
 
 // Whack everything in bss by setting it all to zero. Each value is
 // x + 0x100*y
-int snake[SNAKE_SIZE + 1] = {0};
+short snake[SNAKE_SIZE + 1] = {0};
 short food = 0;
 int length = 0;
